python-backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-telegram-bot==20.7
requests==2.31.0
sqlite3
python-dotenv==1.0.0
cryptography==41.0.4
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==1.1.3
https-proxy-agent==1.0.0